Changed Sprite Sizes and Added Tiles
The recommended pixels per unit for a sprite is the dimensions of said sprite. Therefore, I changed all units from 1 (as the base unit) to 16 (as the new base unit).

Also, adding sprites to a palette creates Tiles (as in the object Unity recognizes). Each terrain sprite has its corresponding Tile object.